refactor: extract login and register schemas into separate files
